refactor: kafka는 kafka 폴더에서 관리
