+++ src/main/java/com/myteam/chat/kafka/service/KafkaProducerService.java
@@ -1,4 +1,4 @@
-package com.myteam.chat.chat.service;
+package com.myteam.chat.kafka.service;
 
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
